CODE RUB: Remove Intern
