+++ InternTrack.Core.Api/Brokers/Storages/IStorageBroker.Interns.cs
@@ -11,5 +11,6 @@ namespace InternTrack.Core.Api.Brokers.Storages
 {
     public partial interface IStorageBroker
     {
+        ValueTask<Intern> DeleteInternAsync(Intern intern);
     }
 }
